+++ src/gl/pipeline/passes/OutlinePass.cpp
@@ -3,8 +3,8 @@
 #include "nfx/graphics/gl/core/Context.h"
 #include "nfx/graphics/gl/core/GlDefinitions.h"
 #include "nfx/graphics/gl/mesh/Mesh.h"
+#include "nfx/graphics/gl/pipeline/raster/ViewportRect.h"
 #include "nfx/graphics/gl/pipeline/RenderState.h"
-#include "nfx/graphics/gl/pipeline/ViewportRect.h"
 #include "internal/runtime/Error.h"
 
 #include <embedded_shaders.h>