+++ src/gl/pipeline/passes/Polygon2DPass.cpp
@@ -2,8 +2,8 @@
 
 #include "nfx/graphics/gl/core/buffers/VertexLayout.h"
 #include "nfx/graphics/gl/core/Context.h"
+#include "nfx/graphics/gl/pipeline/raster/ViewportRect.h"
 #include "nfx/graphics/gl/pipeline/RenderState.h"
-#include "nfx/graphics/gl/pipeline/ViewportRect.h"
 #include "internal/runtime/Error.h"
 
 #include <embedded_shaders.h>